registration page component done
