+++ src/pages/loginPage.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import Title from '../components/Title';
-import LoginForm from '../components/LoginForm';
+// import LoginForm from '../components/LoginForm';
 
 const LoginPage = () => {
   return (
     <>
       <Title text={<span className="titleApp">Finance App</span>} />
-      <LoginForm />
+      {/* <LoginForm /> */}
     </>
   );
 };
